+++ Scripts/Services/Conversation-Experimental/Conversation-Experimental.cs
@@ -24,13 +24,13 @@ using IBM.Watson.DeveloperCloud.Utilities;
 using IBM.Watson.DeveloperCloud.Connection;
 using IBM.Watson.DeveloperCloud.Logging;
 
-namespace IBM.Watson.DeveloperCloud.Services.Conversation.v1
+namespace IBM.Watson.DeveloperCloud.Services.ConversationExperimental.v1
 {
 	/// <summary>
 	/// This class wraps the Watson Conversation service. 
 	/// <a href="http://www.ibm.com/smarterplanet/us/en/ibmwatson/developercloud/conversation.html">Conversation Service</a>
 	/// </summary>
-	public class Conversation : IWatsonService
+	public class ConversationExperimental : IWatsonService
 	{
 		#region Public Types
 		/// <summary>
@@ -54,66 +54,10 @@ namespace IBM.Watson.DeveloperCloud.Services.Conversation.v1
 		#endregion
 
 		#region Private Data
-		private const string SERVICE_ID = "ConversationV1";
+		private const string SERVICE_ID = "ConversationExperimentalV1";
 		private static fsSerializer sm_Serializer = new fsSerializer();
 		#endregion
-        /*
-		#region Workspaces
-		/// <summary>
-		/// Gets the available workspaces for the Conversation service
-		/// </summary>
-		/// <returns>Returns true if request has been sent.</returns>
-		/// <param name="callback">Callback.</param>
-		public bool GetWorkspaces(OnGetWorkspaces callback)
-		{
-			if(callback == null)
-				throw new ArgumentNullException("callback");
-
-			RESTConnector connector = RESTConnector.GetConnector(SERVICE_ID, "/v2/rest/workspaces");
-			if(connector == null)
-				return false;
-
-			GetWorkspacesReq req = new GetWorkspacesReq();
-			req.Callback = callback;
-			req.OnResponse = OnGetWorkspacesResp;
-
-			return connector.Send(req);
-		}
-
-		private class GetWorkspacesReq : RESTConnector.Request
-		{
-			public OnGetWorkspaces Callback { get; set; }
-		}
-
-		private void OnGetWorkspacesResp(RESTConnector.Request req, RESTConnector.Response resp)
-		{
-			DataModels.Workspaces workspaces= new DataModels.Workspaces();
-			if (resp.Success)
-			{
-				try
-				{
-					fsData data = null;
-					fsResult r = fsJsonParser.Parse(Encoding.UTF8.GetString(resp.Data), out data);
-					if (!r.Succeeded)
-						throw new WatsonException(r.FormattedMessages);
-
-					object obj = workspaces;
-					r = sm_Serializer.TryDeserialize(data, obj.GetType(), ref obj);
-					if (!r.Succeeded)
-						throw new WatsonException(r.FormattedMessages);
-				}
-				catch (Exception e)
-				{
-					Log.Error("Conversation", "GetWorkspaces Exception: {0}", e.ToString());
-					resp.Success = false;
-				}
-			}
-
-			if (((GetWorkspacesReq)req).Callback != null)
-				((GetWorkspacesReq)req).Callback(resp.Success ? workspaces : null);
-		}
-		#endregion
-           */
+        
 		#region Message
         private const string SERVICE_MESSAGE = "/v1/workspaces";
 		/// <summary>
@@ -184,15 +128,6 @@ namespace IBM.Watson.DeveloperCloud.Services.Conversation.v1
 		}
 		#endregion
 
-		#region Intents
-		#endregion
-
-		#region Entities
-		#endregion
-
-		#region Dialog Nodes
-		#endregion
-
 		#region IWatsonService implementation
 
 		public string GetServiceID()
@@ -215,11 +150,11 @@ namespace IBM.Watson.DeveloperCloud.Services.Conversation.v1
 
         private class CheckServiceStatus
         {
-            private Conversation m_Service = null;
+            private ConversationExperimental m_Service = null;
             private ServiceStatus m_Callback = null;
             private int m_ConversationCount = 0;
 
-            public CheckServiceStatus(Conversation service, ServiceStatus callback)
+            public CheckServiceStatus(ConversationExperimental service, ServiceStatus callback)
             {
                 m_Service = service;
                 m_Callback = callback;
@@ -240,7 +175,7 @@ namespace IBM.Watson.DeveloperCloud.Services.Conversation.v1
                     OnFailure("Please define a workspace variable in config.json (" + SERVICE_ID + "_ID)");
                 }
             }
-           
+
             private void OnMessage(MessageResponse resp)
             {
                 if (m_ConversationCount > 0)